Remove prints causing exeption (To be fix) Separate USB vitual Com
