+++ usb/include/usb_misc.h
@@ -54,6 +54,7 @@ extern int DbgConsole_Printf(const char *fmt_s, ...);
 
 #if defined(SDK_DEBUGCONSOLE) && (SDK_DEBUGCONSOLE < 1)
 #define usb_echo printf
+
 #else
 #define usb_echo DbgConsole_Printf
 #endif
