+++ usb_app/usb_app_termmsg.h
@@ -1,0 +1,14 @@
+/*
+ * usb_app_termmsg.h
+ *
+ *  Created on: 05/10/2019
+ *      Author: uid87057
+ */
+
+#ifndef USB_APP_TERMMSG_H_
+#define USB_APP_TERMMSG_H_
+
+const char WelcomeMsg[64]= {'W','e','l','l','c','o','m','e',' ','t','o',' ','U','S','B',' ','C','D','C',' ','s','e','r','i','a','l',' ','d','e','m','o','\n','p','r','e','s','s',' ','k','e','y',' ','S',' ','t','o',' ','s','t','a','r','t','\n'};
+
+
+#endif /* USB_APP_TERMMSG_H_ */
