+++ usb_sources/usb_device_config.h
@@ -130,7 +130,7 @@
 #endif
 
 /*! @brief Whether the keep alive feature enabled. */
-#define USB_DEVICE_CONFIG_KEEP_ALIVE_MODE (0U)
+#define USB_DEVICE_CONFIG_KEEP_ALIVE_MODE (1U)
 
 /*! @brief Whether the transfer buffer is cache-enabled or not. */
 #define USB_DEVICE_CONFIG_BUFFER_PROPERTY_CACHEABLE (0U)
